rename accept_blocks.rb to block.rb
