+++ lib/uninclude.rb
@@ -1,3 +1,3 @@
 require 'uninclude/version'
 require 'uninclude/uninclude'
-require 'uninclude/accept_blocks'
+require 'uninclude/block'
