+++ lib/uninclude/block.rb
